Add jobtap plugin for datawarp string validation
Problem: datawarp strings are opaque to Flux and require validation by
Datawarp Services before a job can be fully queued.

Solution: add a jobtap plugin that automatically inserts a dependency
for jobs that contain `attributes.system.dw` in their jobspec.  Send an
RPC for validation of the string, and only remove the dependency if the
validation was successful.  If validation fails, raise a job exception.