cray-libpals: add port-distribution jobtap plugin
Each job using Cray's MPI needs a set of ports that are
exclusive to the set of nodes the job is running on. As a
first approximation, add a jobtap plugin for distributing ports
globally.

The ports distributed can be configured with port-min and
port-max.

Ports are only distributed if the job has more than one
job shell. Ports are only needed if the job takes
multiple nodes, but multiple shells makes for
easier testing.

Closes #6