modules/dws: support creating `Workflow` objects
Problem: DWS expects the WLM to create a `Workflow` object for each job
that includes a DW string.

Solution: Add a python script that registers the `dws` service and
handles the `dws.create` RPC. Handling the RPC constitutes create a
Workflow CRD instance via the K8s API and checking the response for
errors, and then responding to the jobtap RPC with the success/failure
of the Workflow creation.

Update the jobtap plugin that generates the `dws.create` RPC to include
the `jobid` in the RPC payload, which is a field required by the
`Workflow` object.

Add a test that checks the `Workflow` creation works with valid DW
strings and fails with invalid DW strings.

Add/modify sharness.d scripts to support finding the K8s credentials
and the new python modules under sharness.