test: run tests serially to avoid conflicts in K8s
Problem: some of the tests interact with a single, shared K8s cluster.
While the tests as a whole are designed to be idempotent, their
interactions with K8s are not atomic.

Solution: Run the testsuite serially (e.g., `make check`) as opposed to
in parallel (e.g., `make -j check`) to avoid tests conflicting with one
another's expected K8s state.