dws: add shell plugin to set job environment
When a Workflow completes the pre-run state, DWS assigns
it a few environment variables which must be set in order
for the job to use the file systems it requested. Add
infrastructure for setting the variables on the job: send
the variables from dws.py to dws-jobtap.c, and then post
the variables to the job's eventlog, where they can be
read in by a job shell plugin.

Refactor some eventlog-parsing functions out of cray_pals.c
and into a common file which can be used by multiple shell
plugins.