dws: pass updated jobspec to scheduler
Problem: once the dws.py copy of the jobspec has been
updated by to reflect the Rabbit resources needed,
the jobspec copy sent to the scheduler needs to be
updated as well.

Solution: send the dws.py jobspec back to the jobtap
plugin to be updated before submission to the scheduler.