DWS: handle job exceptions
When a job hits an exception, notify DWS and perform cleanup actions.
If the exception occurs while the job is running, nothing special
needs to be done--treat like normal nonzero termination. But if
the exception occurs earlier, remove the dws-jobtap prolog
(if present) and transition the associated Workflow CRD to
'teardown' immediately.