dws: handle job exceptions
Problem: dws.py does not handle job exceptions, for
instance from a user-requested cancellation.

Solution: when a job hits an exception, notify dws.py and
perform cleanup actions. If the exception occurs
while the job is running, nothing special
needs to be done--treat like normal nonzero termination. If
the exception occurs earlier, remove the dws-jobtap prolog
(if present) and transition the associated Workflow to
'teardown' immediately.