cray-pals: obtain ports under prolog action in jobtap plugin
Problem: Users of the cray_pals_port_distributor jobtap plugin
(e.g. the corresponding shell plugin) have no way to determine if
the jobtap plugin is loaded because the 'cray_port_distribution'
event is emitted asynchronously to the job's eventlog.

Switch the jobtap plugin to emit the event under a prolog action. This
delays the start request for the job until the cray_port_distribution
event has been emitted, and therefore will allow the job shell plugin
to determine that the cray jobtap plugin is not loaded if the event
is not posted before the 'start' event.