cray: detect missing jobtap plugin in shell plugin
Problem: The cray libpals shell plugin blocks for 10s waiting for
the cray_port_distribution event, then exits with an error if it
is not found, but this is expected behavior when the corresponding
jobtap plugin is not loaded. The shell plugin should instead disable
(part of) its operation when the jobtap plugin is missing.

Now that the jobtap plugin emits the cray_port_distribution event
under a prolog action, the shell plugin can assume the jobtap plugin
is not loaded when the job "start" event is received before the cray
event. In this case continue without setting PMI_CONTROL_PORT in the
job's environment. This avoids unnecessary fatal job exceptions when
the jobtap plugin is not loaded, and the shell plugin will only block
until the start event is posted.