dws: fix server-updating logic
Problem: DWS Servers require an 'allocationCount' value equal
to the number of compute nodes allocated with a given rabbit.
Prior, dws.py treated allocationSize and allocationCount as
combining to make a single mass.

Solution: make dws.py keep track of which compute nodes are
associated with which rabbits, and give DWS servers a precise
allocationCount value.