cray-pals: fix incorrect task counts
Problem: the libpals shell plugin does not account for task counts
not specified in the usual task-count jobspec position. For instance,
it does not count tasks specified with the per-resource.* shell options.

Solution: properly count tasks by checking the counts given by the shell
rather than the counts given in the jobspec.

Fixes #27.