dws: watch: handle expired resource versions
Problem: when watching resources, if the only existing objects
are old and have not been updated in some number of minutes,
attempting to watch all changes since their last update will
cause Kubernetes to throw an error, because those changes
no longer exist in memory.

Solution: following Kubernetes docs, re-issue a watch request
with a resource version of 0, to get the latest changes that
are available.