dws: choose rabbits based on compute nodes
Problem: while figuring out a permanent solution
for use in production, Flux needs a way to pick
rabbits to satisfy user requests for rabbit
storage. The current implementation is broken
and requires patches to work on specific systems.

Solution: for now, keep the scheduler ignorant of the
rabbits, and make scheduling decisions based on normal
factors like cores and nodes. Then, pick rabbits based
on the set of compute nodes chosen.