dws: split DW directives by newline
Problem: in order to specify multiple #DW
directives on the command line or in a batch script,
there must be a way for a single string to contain
multiple directives.

Solution: split the system.dw jobspec attribute
on newlines. As long as each #DW directive is on its
own line, multiple directives should be supported.