rc: add coral2 rc script
Problem: the cray_pals_port_distributor jobtap plugin is not loaded
in all Flux instances.

Instead of adding the jobtap plugin to the system config which
only applies to the Flux system instance, provide an rc fragment
which loads the plugin in all Flux instances.

If the plugin is already loaded, print a warning message but
continue on, to facilitate transition from the config method.

Fixes #52