dws: make coral2_dws jobid checks optional
Problem: coral2_dws.py expects to be run as a flux job
and expects to find FLUX_JOB_ID in its environment. But
it is being transitioned to becoming a systemd service.

Make checks for FLUX_JOB_ID optional. They are still
useful for testing purposes, so do not removed them
completely.