build: require flux-core >= 0.49.0
Problem: the cray-pals shell plugin parses the pmi shell option
as a list of PMI implementations and if it doesn't find itself
in the list, it is disabled. However, in flux-core versions prior
to 0.49.0, the pmi shell plugin throws an error if the pmi shell
option does not contain options it recognizes.

Require flux-core >= 0.49.0 at configure time.

Fixes #65