dws: stop saving rpcs for response
Problem: the coral2-dws service saves incoming RPCs so it
can respond to them later. However, this means that if the
service crashes, it can no longer communicate with the
jobtap plugin about jobs that were in flight.

Stop saving incoming RPCs and instead of responding
to incoming RPCs, send new RPCs instead.