dws: fetch jobspec resources field dynamically
Problem: the coral2_dws service needs to modify the
'resources' section of a jobspec. It currently relies
on being sent the resources in the workflow creation RPC.

However, in the event that the service restarts, the
service won't have the resources section of any
outstanding jobs.

Add logic to fetch the resources section of jobs on demand.