dws: add workflow creation tracking
Problem: the dws-jobtap plugin does not track whether
a workflow was successfully created for a job. This can
cause issues when the plugin tries to tell the coral2_dws
service to delete a workflow, when in fact no workflow
was ever created.

Before sending a workflow deletion request, check that
a workflow was created to begin with, by looking for a
boolean aux set after creation.