dws: update jobspec manipulation logic
Problem: the directivebreakdown-to-jobspec
conversion functions in directivebreakdown.py
are broken and unused. The coral2-dws.py
script feeds DWS back the requirements given
in the directivebreakdowns, it does not
meaningfully check them.

As a first-cut solution, change users'
jobspecs to add a top-level 'rack' entry,
with 'node' and 'rabbit' children.
Place all rabbit storage requests under
that 'rabbit' vertex.