dws: change Error to TransientCondition
Problem: HPE has changed their error handling for workflows.
Previously, all problems were reported as 'status: Error' and it was
impossible to distinguish transient issues from fatal errors.
Now fatal errors are reported as 'status: Error' and issues that are
not known to be fatal are reported as 'TransientCondition'.

If a workflow enters 'TransientCondition' state, add it to a data
structure and throw an exception if it does not leave that state
within a certain time interval.