dws: add checks for workflow Error state
Problem: HPE has changed their error handling for workflows.
Previously, all problems were reported as 'status: Error' and it was
impossible to distinguish transient issues from fatal errors.
Now fatal errors are reported as 'status: Error' and issues that are
not known to be fatal are reported as 'TransientCondition'.

If a workflow enters the 'Error' state, raise a job exception.