dws: add --resourcegraph option to coral2_dws.py
Problem: the coral2_dws service needs to be able to mark rabbits
as up or down, so that Fluxion will not schedule jobs on down
rabbits.

In order to mark rabbits as down, the service must know the path
of each rabbit in Fluxion's resource graph.

Add a command-line argument which provides the path to Fluxion's
resource graph.