dws: send rpcs to mark rabbits as up or down
Problem: the coral2_dws service needs to be able to mark rabbits
as up or down, so that Fluxion will not schedule jobs on down
rabbits.

Rabbit status is reported by 'Storage' resources in k8s.

Watch Storage resources for changes and mark rabbits as up or down
when a change occurs by sending an RPC to the scheduler, fetching
the path of each rabbit in Fluxion's resource graph from the JFG
representation stored in a file.