cray-pals: allow for duplicate nodes for testing
Problem: it is difficult to test the cray-pals plugin with multiple
nodes because it uses a unique hostlist to determine the nodecount.
When running with multiple brokers per node (such as with
`--test-size=2`), the other information it fetches from the shell,
like the task mapping, treat the duplicated nodes as distinct.

Make the plugin use a non-unique hostlist to allow duplicate nodes.