dws: reduce lustre allocation fragmentation
Problem: lustre allocations are split across rabbits based on the
number of compute nodes associated with each rabbit, but this
sometimes results in undue fragmentation. DWS requires that rabbit
resources be specified as a top-level size and then a number of
allocations of that size to create for each rabbit.

Given six nodes on one rack and six on another, the coral2_dws
plugin would ask for six allocations on one rack and six on the
other, with each allocation being 1/12 the total size; it would
be more efficient to create one allocation on each rack with each
allocation being 1/2 the total size.

Fix this by checking for the greatest common divisor of the
nodecount on each rack, and scaling up the allocation size and
scaling down the allocation count accordingly.

For cases like seven nodes on one rack and nine on the other,
(i.e. the nodecounts have a gcd of 1) there is no improvement.