dws2jgf: support rabbits as computes
Problem: the dws2jgf script does not support rabbit nodes as compute
nodes. At the moment they would be treated as compute nodes
independent of any rack or rabbit.

Adding them as computes under rack vertices would cause a collision
in the generated resource graph: there would be two vertices with
the same path, one with type "node" and the other with type
"rabbit".

If rabbit nodes are found in R, add them under rack vertices with
the prefix 'rabbit-' to distinguish them in the resource graph
from the compute node vertex.