test: improve rabbit-watching tests
Problem: the tests for watching rabbit status rely on checking logs
for specific messages. This is a difficult and unreliable form of
testing.

Add more thorough tests checking that rabbits are in fact marked as
down by Fluxion--when the corresponding K8s Storage object is marked
as down--by submitting rabbit resource requests.