test: fix race condition in t1000-dws-depend
Problem: there is a race condition in t1000-dws-dependencies.t.
It imitates a failure response to the dws.post_run RPC but then
sends a separate success RPC afterwards, confusing the jobtap
plugin and creating a race between the RPCs, which should never
happen under normal circumstances.

Prevent the race condition by not sending the success RPC in the
failure case.