dws-jobtap: fix race condition in exceptions
Problem: the 'raise_job_exception' helper function invokes
'flux_job_raise' which is an RPC, so the exception is not guaranteed
to have been raised when the function returns. Exceptions should be
raised before the prolog and epilog actions are removed.

Use 'flux_jobtap_raise_exception' instead, and remove the prolog
and epilog actions only after the exception has been raised.