dws: add a special exit code for k8s problems
Problem: when the k8s server runs into problems, the coral2-dws
enters a restart loop. This clutters the logs.

Add a special exit code for kubernetes problems so that systemd
can stop restarting the service when it exits with that code.

Use 3 for the exit code since 1 is for general issues and 2 is for
command-line syntax errors.