systemd: recognize special flux-dws exit code
Problem: systemd should recognize a special exit code for kubernetes
problems and halt its restarts of the service so that logs are not
cluttered while admins resolve the issue.

Systemd should also recognize a special exit code for command-line
syntax issues, since the unit file would need to be altered.

Add a 'RestartPreventExitStatus' option to the systemd unit file for
the service.

Fixes #144.