dws: handle k8s resourceversion exceptions
Problem: while watching for k8s resources, a resourceversion > 0
is required. When there are only old resources lying around and
no new ones, the resourceversion will be > 0 but too old for
kubernetes to accept it, and an exception may be raised.

Catch the exception and set the resource version to 0.