dws2jgf: drop rabbit vertices from graph
Problem: in the current way flux-dws2jgf creates resource graphs,
every rack vertex has a single rabbit child vertex, which in turn
has numerous ssd children. This makes it difficult for Fluxion to
schedule when multiple nodes are needed on a single rack.

Remove the 'rabbit' vertices and make the ssd vertices the direct
children of 'rack'.