test: update tests to use alloc-bypass
Problem: some of the rabbit jobspecs have a top-level
'slot' entry with 'node' below that, forcing the node to be
allocated exclusively. In testing on a single-node system, that
can cause a deadlock because the coral2_dws script is occupying
resources that another job needs, but the job will fail if
coral2_dws isn't running.

Use alloc-bypass to execute coral2_dws to work around the issue.