dws: update coral2-dws to mark down ssds
Problem: when a k8s storage resource isn't in the "Ready" state,
the coral2-dws service attempts to send an RPC to Fluxion to mark
the associated rabbit vertex as offline. Now that there are no
rabbit vertices, that RPC will fail.

Change the RPC to mark all ssds as offline.