dws2jgf: add ssds per rack as a rack property
Problem: in order to mark rabbit storage as offline, coral2-dws will
have to mark every ssd under 'rack' as offline. To do that, it will
need to know the number of ssd vertices associated with each rack.

Coral2-dws could read the entire graph to discover that, but that
would require more parsing than desirable.

Add a property to rack vertices giving the count of ssd children.