dws: add flag to disable fluxion scheduling
Problem: Fluxion scheduling of rabbits is unfortunately still in
flux. Particularly on elcap, where we have decided not to roll out
the JGF representation of the full resource graph while other
problems are worked out.

Add a command-line argument to disable features that require Fluxion
to have the complete JGF representation of the graph, as generated
by dws2jgf.