dws2jgf: change cluster-name logic
Problem: at the moment admins should supply the name of the cluster
through a command-line argument when running dws2jgf. But they will
probably run it from the management node of a cluster, and most LC
clusters have nodes with names like 'nameXXXXX' for XXXXX some
decimal number.

Simplify things for them by making the cluster name in the resource
graph default to the hostname stripped of numerals, with trailing
dashes stripped as well.