dws: fix PostRun workflow Error hang
Problem: if a workflow goes to status: Error, coral2_dws will raise
an exception. For most states the exception will transition the job
to CLEANUP, which eventually moves the workflow to Teardown.
However, if the job is already in CLEANUP and is being held in the
dws-epilog epilog when the exception is thrown, the exception has
no effect and the epilog will never be finished because the workflow
does not progress.

If a workflow moves to status: Error in PostRun or DataOut states,
move it directly to Teardown.