dws: add nnf-dm boolean to resource-update RPC
Problem: There is no marker that a job needs the nnf data movement
daemon to be running during its duration. Without a marker, an
administrative prolog cannot tell whether to start the daemon.

Add a boolean to the DWS resource-update event that will tell the
jobtap plugin (which receives the RPC) whether or not the daemon
is needed.