dws-jobtap: add copy-offload flag to env event
Problem: There is no marker that a job needs the nnf data movement
daemon to be running during its duration. Without a marker, an
administrative prolog cannot tell whether to start the daemon.

Add a boolean to the dws-environment event that indicates whether
the nnf-dm daemon is needed. The administrative prolog can watch for
this event and its context.

Fixes #167.