dws: move workflows in TC to teardown
Problem: as in #169, for most states, raising an exception should
be enough to trigger other logic that eventually moves the workflow
to Teardown. However, if the workflow is in PostRun or DataOut, the
exception won't affect the dws-epilog action holding the job, so
the workflow should be moved to Teardown immediately.

Move workflows that are stuck in TransientCondition in DataOut or
PostRun to Teardown immediately.