dws: first pass at handling Lustre MGTs
Problem: Flux does not handle workflows that require Lustre MGTs,
or more specifically any DirectiveBreakdown with an allocation type
of 'AllocateSingleServer'. However, Flux handles MDTs and other
allocations which have a 'count' constraint of 1, which is the same
thing effectively.

Handle 'AllocateSingleServer' allocations the same way that 'count'
constraints are handled, which is to pick a single rabbit and
place the allocation there, ignoring Fluxion's allocation choices
somewhat.