dws: improve dws_environment shell plugin errors
Problem: the dws_environment shell plugin's errors are less clear
than they could be, and do not distinguish the case where
the 'dws_environment' event is not posted from the case where the
event_watch is simply slow.

Add a check for the 'start' event and error out if it is found
before the 'dws_environment' event.