dws-jobtap: raise job exceptions for size limits
Problem: the coral2_dws script sends a resource-update RPC with a
non-null 'errmsg' field if a job requests storage that is beyond the
size constraints for file systems specified on the coral2_dws
command line. However, the dws-jobtap plugin does not act on those
error messages.

Add logic to the jobtap plugin to raise an exception if the errmsg
field is not null.