dws2jgf: add script for generating rabbit map
Problem: as described in #193, JGF is too unwieldy to be stored in
Ansible. On the other hand, Flux's ability to start up and run
jobs cannot be dependent on the responsiveness of kubernetes, so
generating JGF from kubernetes before starting Flux is not an
option.

A solution would be to store some static rabbit data in ansible,
generated by reading from kubernetes. This data could be read in
to generate JGF.

Add a script that generates a JSON file describing which nodes map
to which rabbits and what the capacity of each rabbit is.