dws2jgf: read from JSON file instead of k8s
Problem: as described in #193, JGF is too unwieldy to be stored in
Ansible. On the other hand, Flux's ability to start up and run
jobs cannot be dependent on the responsiveness of kubernetes, so
generating JGF from kubernetes before starting Flux is not an
option.

Change flux-dws2jgf to read from a static JSON file generated by
the flux-rabbitmapping script, instead of from kubernetes.