test: make flux-sharness.sh current with core
Problem: flux-sharness.sh is out of date with flux-core, where it
came from. Also, there are HAVE_JQ prereqs on many test cases, but
flux-core requires JQ for all tests, so the testsuite for this repo
could have the same requirement.

Update to require JQ for the entire testsuite by pulling in some
changes to flux-sharness.sh from flux-core.