dws: ignore bad jobids in workflows
Problem: sometimes users create workflows with nonsense jobids.
For instance, on Tuolumne recently a user created a workflow with
jobid "manual", crashing the service because the jobid could not
be converted to a flux job ID.

Catch all exceptions that occur while trying to fetch basic data
about a workflow and log them, but do not let them crash the
service.