dws: make coral2_dws read systemconfiguration
Problem: coral2_dws lists `Storage` resources to get the mapping of
compute nodes to rabbits, but some `Storage` resources may not be up.
Since the mapping is only initialized when the coral2_dws script
starts, errors can occur if `Storage` resources come online later
because `coral2_dws` will not recognize which compute nodes that
`Storage` is attached to.

Check the `Systemconfiguration` instead, which is comprehensive and
more reliable.