dws: move command-line options to config
Problem: a number of rabbit configuration options, such as the
number of nnfdatamovement resources to save and whether or not
to restrict the creation of persistent file systems to the
instance owner, are set by command-line options to the coral2-dws
service.

Move those parameters to a config file and drop the command-line
options.