dws: support toggling rabbit draining in config
Problem: the option to disable the draining of compute nodes that
lose connection with their local rabbit is set by a command-line
option to the coral2-dws service. It should a config file option.

Move the draining option to a config file and make coral2-dws read
from it. Drop the command-line option.