dws: drop resourcegraph command-line option
Problem: the coral2_dws script has a command-line option to set the
path of the file from which it reads JGF for the rabbit resources
it will operate. However, that is unnecessary and unhelpful because
it could instead fetch the `resource.R` KVS key, which is a more
reliable option.

Drop the command-line argument and make the script read from the
KVS.