dws: fix broken TransientCondition message
Problem: workflows' TransientCondition message is not transmitted to
users correctly. Users get 'None' as an explanation for what the
problem was, because the actual error string is set to a property
named 'last_tc_message' but the property read from is named
'last_message' (and it is initialized to None).

Set the error string to the 'last_message' property which will make
the error message be transmitted to users properly.