dws: track hostlist of nodes with down rabbits
Problem: Fluxion continues to have problems with rabbits on
production systems. However, users need a way for Flux to
automatically schedule around down rabbits.

As a workaround, maintain a hostlist of all nodes that are attached
to down rabbits. Eventually this hostlist should be added to jobs'
constraints, as a negative hostlist constraint (like
--requires=-hosts:foobar[1-20])