dws-jobtap: use jobspec_update_id for resources
Problem: The dws-jobtap plugin does not use the jobspec update
jobtap interface to perform its update of a job's resources. Instead
it simply grabs the jobspec directly and manipulates it. However, it
would be good for users to have a record of how their jobspec was
changed, and doing so may prevent problems in the future.

Make dws-jobtap use the jobspec update interface.

Fixes #158