dws: add hostlist constraint to rabbit jobs
Problem: while Fluxion scheduling of rabbits continues to have
problems in production, a stopgap is needed to prevent rabbit jobs
from running on nodes attached to down rabbits.

Make the dws-jobtap plugin add a hostlist constraint for rabbit jobs
based on the hostlist of nodes with down rabbits that coral2_dws
sends.