dws-jobtap: send teardown rpc on epilog exception
Problem: jobs can get stuck in epilog if a workflow gets stuck in
the DataOut stage, and users currently have no way to force the
workflow to move to Teardown.

Make it so that if an exception event is received while the
dws-epilog action is running, a dws_teardown rpc is sent, which
should move the workflow to Teardown and lead to the release of the
epilog action.