dws: add badrabbit property if PCIe link fails
Problem: in theory, if a rabbit-to-compute-node PCIe link goes down,
coral2_dws will drain the node. However, admins have generally
disabled that feature for being disruptive to non-rabbit users.
PR #265 introduced a non-intrusive way to avoid scheduling rabbit
jobs on specific nodes, by tagging nodes with a property. This
could be used to cover the PCIe link failure case as well.

If draining is disabled, unless the `rabbit.soft_drain` config
option is set to `false`, mark nodes with a failed PCIe link
with the `badrabbit` property.