dws: don't always remove badrabbit property
Problem: with certain configurations, if a rabbit is down, coral2_dws
sets the 'badrabbit' property on all the compute nodes attached to
it. It then proceeds to go through each compute node individually and
set or remove the 'badrabbit' property depending on whether that
node's PCIe link is up or down. However, if the rabbit as a whole is
down, removing the property is incorrect--the property should stay
there.

Add a check to make sure the rabbit as a whole is ready before
removing the property.