`getrabbit`: jobids lookup rabbits in memo
Problem: PR #272 added a way to look up the rabbits for a job.
It currently works correctly in all cases, however in the future for
lustre file systems, the rabbits associated with a job may no longer
be the rabbits local to the compute nodes in the job.

There is already a memo associated with every rabbit job, specifying
what rabbits are used. Use that instead.

Fixes #275.