dws: refactor `badrabbit` property logic
Problem: there are three different kinds of handling for down
rabbits, and the code for them is messy.

First, there is draining compute nodes. Second, there is marking
rabbit vertices down in Fluxion. Lastly and most recently added,
there is the setting of the `badrabbit` property on nodes.

The third method is the most complicated and the code implementing
it is mixed in with the code for the other two types of handling.
It should be broken out to be handled separately and more cleanly.

Break out the property logic into its own function.