dws-jobtap: avoid setting aux too early
Problem: as described in #285, the `dws-jobtap` plugin creates a
boolean jobtap aux to indicate whether or not the prolog it sets
on jobs is still active, but it creates the aux in the DEPEND state
despite the fact that it will not start the prolog until the RUN
state. If Flux is restarted with jobs in the queue, the aux
will be lost and a lookup for the aux will cause the RUN callback
to fail.

Only set the `prolog_active` aux at the time that the prolog is
actually started, and not before.

Fixes #285.