dws: allow RPC error responses without exceptions
Problem: the dws.create RPC sometimes rejects jobs due to failing
checks. However, whenever it does so it raises an exception and logs
a traceback, which is fine but is unnecessary, since the problems
are user error and should not be logged.

Add special exception handling for user errors that does not result
in logging.