dws-jobtap: stop tracking workflow creation
Problem: dws-jobtap.c tracks whether a workflow has been created for
a job by setting an aux. If the aux is not set, it does not send
a `dws.post_run` RPC, to avoid causing errors with `coral2_dws`
if that script did not find a workflow. However, `coral2_dws` can
now handle missing workflows. Besides, the aux could be lost if the
plugin was reloaded, and that would cause errors.

Stop tracking workflow creation with an aux.