libapinfo: add convenience library
Problem: HPE apinfo file support is embedded in the cray_pals
shell plugin, which makes it hard to follow, awkward to develop
unit tests for, and harder for HPE to help us support.

Create a libapinfo convenience library to make the apinfo bits
self contained and separately testable.

The new library is plumbed to support multiple apinfo versions,
as we are currently on an old one and may want to be able to revert
back to it as we move forward.