dws-jobtap: check state before dependency-remove
Problem: the dws-jobtap plugin may receive multiple dependency-remove
RPCs from the coral2_dws script. However, if it does, it currently
raises an exception on the target job. This is not desirable for
users.

Upon receipt of the RPC, check if the job is in DEPEND, and if not,
return an error response. If the dependency removal fails, do not
raise an exception but log an error.