dws: handle duplicate directivebreakdown update
Problem: `directivebreakdown.py` expects a jobspec's resources to
be laid out a certain way, with a top-level 'node' entry. It then
modifies the resources to add rabbit resources, and once that is
done, the same code may reject the jobspec if it were passed through
again. However, that can happen if the `coral2_dws` service is
restarted with fluxion scheduling enabled.

If fluxion scheduling for rabbits is disabled, this problem does not
apply because the jobspec update is not applied.

Add an early `return` statement to `apply_breakdowns`, that checks
whether it looks like the jobspec has already been updated, and if
so returns a special return value.