libapinfo: add v5 support
Problem: the cray-pals plugin uses an old version of the apinfo
file, but the new version will be required for advanced features.

Update to HPE-provided header and more or less copy the apinfo1
implementation to a new apinfo5 one.  Although they are currently
duplicates, this will allow the v5 code to evolve without breaking
the known good v1.

See also #267