dws2jgf: fix properties on nodes not in DWS
Problem: when flux-dws2jgf writes out nodes, it neglects to output
properties for all nodes that are not in DWS. For instance, on Tioga,
tioga[42-43] are not associated with a rabbit, but they are listed in
resource.toml and are assigned to queue mi300a. But flux dws2jgf
wrote them out without assigning any properties whatsoever.

Fix properties.