cray-pals: add global context container
Problem: cray-pals doesn't have a container that can hold
state accumulated as the plugin runs, which increases the
complexity of internal argument passing and makes the plugin
laborious to extend.

Move some things like configuration and fetched shell info
into a context struct that is freed when the plugin is unloaded.

Move the parsing of plugin options to a dedicated function
and improve error handling slightly.