cray-pals: add -ocray-pals.pmi-bootstrap option
Problem: the method of obtaining Cray PMI bootstrap info
(ports, secret) is hardwired to use the cray_port_distribution
jobtap plugin, but other methods need to be explored.

Refactor to allow the bootstrap info to come from other sources.
Add the ability to set it on the command line, for testing.