dws: check for workflow existence in teardown rpc
Problem: the `dws.teardown` RPC handler in `coral2_dws` does not
check whether the workflow exists before trying to move it to
Teardown. However, if a user cancels a job repeatedly shortly after
submission, the workflow may not be created before the `dws.teardown`
RPC is sent.

Make the RPC handler check for workflow existence.