dws: on abort, look for active mounts to drain
Problem: when the 'abort' message is received, Flux should look for
clientmounts that are still active on compute nodes and drain them,
so that the jobtap epilog action can be removed and all of the nodes
that have completed unmounting can be returned to service.

Add logic to look for clientmounts that aren't in the 'unmounted'
state and drain the associated nodes.