dws: on abort, disable all rabbits with allocs
Problem: when the 'abort' message is received, Flux should look for
rabbits that haven't yet cleaned up their allocations and disable
them, so that the jobtap epilog action can be removed and all of
the nodes that have completed unmounting can be returned to service.
The rabbit resources would be returned to service as well if the
rabbit were not disabled, potentially causing problems for future
jobs.

Add logic to look for servers resources that have allocation sizes
greater than 0, and disable the storage resources associated with
them.