flux_k8s: refactor storage module for classes
Problem: the flux_k8s.storage module consists of a lot of complicated
functions that take a lot of arguments. At the same time, it exposes
relatively little functionality outside the module.

Refactor the existing code into two classes, a RabbitManager and a
FluxionRabbitManager subclass, that expose functionality for setting
properties on compute nodes.