dws: catch remove-finalizer exception and retry
Problem: on Hetchy, Flux failed to remove finalizers from a workflow.
The workflow had already been deleted, perhaps manually, and attempts
to remove the finalizer failed because the `remove_finalizer`
function was using an old version of the workflow with finalizers
that had been removed. Therefore, the calls to patch the workflow
to remove the Flux finalizer were seen by kubernetes as an attempt to
*add* finalizers, which is not allowed if the workflow has been
deleted, and a 422 error code was returned.

Add new exception-catching logic to `remove_finalizer`, fetching the
latest workflow and trying again.