dws: support data movement secret
Problem: in order to support the new style of rabbit data movement,
Flux needs to grab a kubernetes secret for certain workflows, and
then export it to jobs as an environment variable.

Grab the secret and pack it into the prolog-remove RPC for it to
be set in the job's environment by the job shell.

Fixes #256.