tests: add test for new rabbit copy-offload token
Problem: there are no tests that Flux properly fetches the k8s secret
used for the new copy-offload API.

Add a test.

To test, a copy-offload container profile must be used. However,
in a kubernetes kind environment, there is no global lustre file
system, so the default container profile for copy-offload must be
modified. This seems to make the workflow hang in PostRun, so cancel
the job.