dws: do not set property on non-compute nodes
Problem: the `flux_k8s.storage` module can send RPCs to mark
properties on nodes that Fluxion does not have in its resource
graph, leading to a lot of warnings in the logs. Sometimes this
alarms admins, although it has no actual consequences.

Only send RPCs for nodes that are part of the Flux instance and are
not listed in the `resource.exclude` config key.

Fixes #296.