dws: fix memory leak in callback
Problem: in `cleanup_cb`, there are multiple `return` statements
after some memory has been allocated but before the memory is
set to a jobtap aux for automatic deletion. Those `return` statements
would therefore cause the memory to be leaked.

Assign the memory to a jobtap aux at the time of allocation.